refactor(code): spill module becomes store
