+++ packages/code/src/store.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, test } from "bun:test"
 import { Effect } from "effect"
 import { KeyValueStore } from "effect/unstable/persistence"
-import { hydrate, refs, spill, WORKSPACE_REFS } from "./spill"
+import { hydrate, refs, spill, WORKSPACE_REFS } from "./store"
 
 // The spill seam over a KeyValueStore: what goes in comes back out whole, and the reserved
 // manifest key indexes the refs one store holds.
